Update PlaceOnTouch.cs
This was originally supposed to be a poster-placing app. Unfortunately, AR in it's current form doesn't recognize walls very well, and vertical planes are difficult to work with. In an effort to pivot, it's now a plant-placing app. Place plants around your apartment!

Today's work was to fix the build and allow for the placement of plants.

I can now tap the screen and have a model show up.

Next, I need to make it so that it's possible for the user to rotate plants, then add functionality to change which plant the user wants to place. Finally, I'll add animations and sound effects and publish my first AR app!